refactor: 관심사에 따른 Group Service 분리
