+++ backend/src/main/java/com/woowacourse/momo/group/service/GroupSearchService.java
@@ -1,0 +1,58 @@
+package com.woowacourse.momo.group.service;
+
+import java.util.List;
+
+import org.springframework.data.domain.Page;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import lombok.RequiredArgsConstructor;
+
+import com.woowacourse.momo.group.domain.Group;
+import com.woowacourse.momo.group.service.request.GroupFindRequest;
+import com.woowacourse.momo.group.service.response.GroupPageResponse;
+import com.woowacourse.momo.group.service.response.GroupResponse;
+import com.woowacourse.momo.group.service.response.GroupResponseAssembler;
+import com.woowacourse.momo.group.service.response.GroupSummaryResponse;
+import com.woowacourse.momo.member.domain.Member;
+import com.woowacourse.momo.member.service.MemberFindService;
+
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class GroupSearchService {
+
+    private final MemberFindService memberFindService;
+    private final GroupFindService groupFindService;
+
+    public GroupResponse findGroup(Long id) {
+        Group group = groupFindService.findGroup(id);
+        return GroupResponseAssembler.groupResponse(group);
+    }
+
+    public GroupPageResponse findGroups(GroupFindRequest request) {
+        Page<Group> groups = groupFindService.findGroups(request);
+        List<Group> groupsOfPage = groups.getContent();
+        List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
+
+        return GroupResponseAssembler.groupPageResponse(summaries, groups.hasNext(), request.getPage());
+    }
+
+    public GroupPageResponse findParticipatedGroups(GroupFindRequest request, Long memberId) {
+        Member member = memberFindService.findMember(memberId);
+        Page<Group> groups = groupFindService.findParticipatedGroups(request, member);
+        List<Group> groupsOfPage = groups.getContent();
+        List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
+
+        return GroupResponseAssembler.groupPageResponse(summaries, groups.hasNext(), request.getPage());
+    }
+
+    public GroupPageResponse findHostedGroups(GroupFindRequest request, Long memberId) {
+        Member member = memberFindService.findMember(memberId);
+        Page<Group> groups = groupFindService.findHostedGroups(request, member);
+        List<Group> groupsOfPage = groups.getContent();
+        List<GroupSummaryResponse> summaries = GroupResponseAssembler.groupSummaryResponses(groupsOfPage);
+
+        return GroupResponseAssembler.groupPageResponse(summaries, groups.hasNext(), request.getPage());
+    }
+}
