+++ backend/src/test/java/com/woowacourse/momo/group/service/GroupManageServiceTest.java
@@ -7,11 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static com.woowacourse.momo.fixture.GroupFixture.DUDU_STUDY;
 import static com.woowacourse.momo.fixture.GroupFixture.MOMO_STUDY;
 
-import java.util.Comparator;
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -25,12 +21,7 @@ import com.woowacourse.momo.category.domain.Category;
 import com.woowacourse.momo.global.exception.exception.MomoException;
 import com.woowacourse.momo.group.domain.Group;
 import com.woowacourse.momo.group.domain.GroupRepository;
-import com.woowacourse.momo.group.service.request.GroupFindRequest;
 import com.woowacourse.momo.group.service.request.GroupRequest;
-import com.woowacourse.momo.group.service.response.GroupPageResponse;
-import com.woowacourse.momo.group.service.response.GroupResponse;
-import com.woowacourse.momo.group.service.response.GroupResponseAssembler;
-import com.woowacourse.momo.group.service.response.GroupSummaryResponse;
 import com.woowacourse.momo.member.domain.Member;
 import com.woowacourse.momo.member.domain.MemberRepository;
 import com.woowacourse.momo.member.domain.Password;
@@ -38,13 +29,13 @@ import com.woowacourse.momo.member.domain.UserId;
 
 @Transactional
 @SpringBootTest
-class GroupServiceTest {
-
-    private static final int PAGE_SIZE = 12;
-    private static final int TWO_PAGE_GROUPS = 8;
+class GroupManageServiceTest {
 
     @Autowired
-    private GroupService groupService;
+    private GroupManageService groupManageService;
+
+    @Autowired
+    private GroupSearchService groupSearchService;
 
     @Autowired
     private GroupRepository groupRepository;
@@ -79,7 +70,7 @@ class GroupServiceTest {
     @Test
     void create() {
         GroupRequest request = MOMO_STUDY.toRequest();
-        long savedGroupId = groupService.create(savedHost.getId(), request)
+        long savedGroupId = groupManageService.create(savedHost.getId(), request)
                 .getGroupId();
 
         Optional<Group> group = groupRepository.findById(savedGroupId);
@@ -116,29 +107,9 @@ class GroupServiceTest {
                 .category(categoryId)
                 .toRequest();
 
-        assertThatThrownBy(() -> groupService.create(savedHost.getId(), request))
+        assertThatThrownBy(() -> groupManageService.create(savedHost.getId(), request))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("존재하지 않는 카테고리입니다.");
-    }
-
-    @DisplayName("모임을 조회한다")
-    @Test
-    void findById() {
-        Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
-        GroupResponse expected = GroupResponseAssembler.groupResponse(savedGroup);
-
-        GroupResponse actual = groupService.findGroup(savedGroup.getId());
-
-        assertThat(actual).usingRecursiveComparison()
-                .isEqualTo(expected);
-    }
-
-    @DisplayName("존재하지 않는 모임을 조회할 수 없다")
-    @Test
-    void findByIdWithNotExistGroupId() {
-        assertThatThrownBy(() -> groupService.findGroup(0L))
-                .isInstanceOf(MomoException.class)
-                .hasMessage("존재하지 않는 모임입니다.");
     }
 
     @DisplayName("모임을 수정한다")
@@ -147,7 +118,7 @@ class GroupServiceTest {
         Group savedGroup = groupRepository.save(MOMO_STUDY.builder().toGroup(savedHost));
         GroupRequest request = DUDU_STUDY.toRequest();
 
-        groupService.update(savedHost.getId(), savedGroup.getId(), request);
+        groupManageService.update(savedHost.getId(), savedGroup.getId(), request);
 
         Optional<Group> group = groupRepository.findById(savedGroup.getId());
         assertThat(group).isPresent();
@@ -180,7 +151,7 @@ class GroupServiceTest {
     void updateNotExistGroup() {
         GroupRequest request = MOMO_STUDY.toRequest();
 
-        assertThatThrownBy(() -> groupService.update(savedHost.getId(), 1000L, request))
+        assertThatThrownBy(() -> groupManageService.update(savedHost.getId(), 1000L, request))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("존재하지 않는 모임입니다.");
     }
@@ -191,7 +162,7 @@ class GroupServiceTest {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
         GroupRequest request = DUDU_STUDY.toRequest();
 
-        assertThatThrownBy(() -> groupService.update(savedMember1.getId(), savedGroup.getId(), request))
+        assertThatThrownBy(() -> groupManageService.update(savedMember1.getId(), savedGroup.getId(), request))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("주최자가 아닌 사람은 모임을 수정하거나 삭제할 수 없습니다.");
     }
@@ -203,7 +174,7 @@ class GroupServiceTest {
         savedGroup.participate(savedMember1);
         GroupRequest request = DUDU_STUDY.toRequest();
 
-        assertThatThrownBy(() -> groupService.update(savedHost.getId(), savedGroup.getId(), request))
+        assertThatThrownBy(() -> groupManageService.update(savedHost.getId(), savedGroup.getId(), request))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("참여자가 존재하는 모임은 수정 및 삭제할 수 없습니다.");
     }
@@ -218,7 +189,7 @@ class GroupServiceTest {
 
         GroupRequest request = DUDU_STUDY.toRequest();
 
-        assertThatThrownBy(() -> groupService.update(savedHost.getId(), groupId, request))
+        assertThatThrownBy(() -> groupManageService.update(savedHost.getId(), groupId, request))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("참여자가 존재하는 모임은 수정 및 삭제할 수 없습니다.");
     }
@@ -228,9 +199,9 @@ class GroupServiceTest {
     void closeEarly() {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
 
-        groupService.closeEarly(savedHost.getId(), savedGroup.getId());
+        groupManageService.closeEarly(savedHost.getId(), savedGroup.getId());
 
-        boolean actual = groupService.findGroup(savedGroup.getId()).isFinished();
+        boolean actual = groupSearchService.findGroup(savedGroup.getId()).isFinished();
         assertThat(actual).isTrue();
     }
 
@@ -239,74 +210,18 @@ class GroupServiceTest {
     void closeEarlyMemberIsNotHost() {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
 
-        assertThatThrownBy(() -> groupService.closeEarly(savedMember1.getId(), savedGroup.getId()))
+        assertThatThrownBy(() -> groupManageService.closeEarly(savedMember1.getId(), savedGroup.getId()))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("주최자가 아닌 사람은 모임을 수정하거나 삭제할 수 없습니다.");
-    }
-
-    @DisplayName("모임 목록중 두번째 페이지를 조회한다")
-    @Test
-    void findAllWithPage() {
-        List<GroupSummaryResponse> summaries = IntStream.range(0, TWO_PAGE_GROUPS)
-                .mapToObj(i -> saveGroup("모모의 스터디", Category.STUDY))
-                .map(GroupResponseAssembler::groupSummaryResponse)
-                .sorted(Comparator.comparing(GroupSummaryResponse::getId).reversed())
-                .collect(Collectors.toList());
-
-        for (int i = 0; i < PAGE_SIZE; i++) {
-            saveGroup("모모의 스터디", Category.STUDY);
-        }
-
-        GroupFindRequest request = new GroupFindRequest();
-        request.setPage(1);
-        GroupPageResponse actual = groupService.findGroups(request);
-
-        assertAll(
-                () -> assertThat(actual.isHasNextPage()).isFalse(),
-                () -> assertThat(actual.getGroups()).usingRecursiveFieldByFieldElementComparator()
-                        .isEqualTo(summaries)
-        );
-    }
-
-    @DisplayName("카테고리의 그룹 목록을 조회한다")
-    @Test
-    void findGroupsByCategory() {
-        for (int i = 0; i < PAGE_SIZE + TWO_PAGE_GROUPS; i++) {
-            saveGroup("모모의 스터디", Category.STUDY);
-            saveGroup("모모의 술파티", Category.DRINK);
-        }
-
-        GroupFindRequest request = new GroupFindRequest();
-        request.setCategory(Category.DRINK.getId());
-        request.setPage(1);
-        GroupPageResponse actual = groupService.findGroups(request);
-
-        assertThat(actual.getGroups()).hasSize(TWO_PAGE_GROUPS);
-    }
-
-    @DisplayName("키워드를 포함하는 이름의 모임을 조회한다")
-    @Test
-    void findAllByKeyword() {
-        saveGroup("모모의 스터디", Category.STUDY);
-        saveGroup("구구의 스터디", Category.STUDY);
-        saveGroup("모모의 술파티", Category.DRINK);
-        saveGroup("두두와의 헬스 클럽", Category.HEALTH);
-
-        GroupFindRequest request = new GroupFindRequest();
-        request.setKeyword("모모");
-        request.setPage(0);
-        GroupPageResponse actual = groupService.findGroups(request);
-
-        assertThat(actual.getGroups()).hasSize(2);
     }
 
     @DisplayName("식별자를 통해 모임을 삭제한다")
     @Test
     void delete() {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
-        groupService.delete(savedHost.getId(), savedGroup.getId());
+        groupManageService.delete(savedHost.getId(), savedGroup.getId());
 
-        assertThatThrownBy(() -> groupService.findGroup(savedGroup.getId()))
+        assertThatThrownBy(() -> groupSearchService.findGroup(savedGroup.getId()))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("존재하지 않는 모임입니다.");
     }
@@ -316,7 +231,7 @@ class GroupServiceTest {
     void deleteNotHost() {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
 
-        assertThatThrownBy(() -> groupService.delete(savedMember1.getId(), savedGroup.getId()))
+        assertThatThrownBy(() -> groupManageService.delete(savedMember1.getId(), savedGroup.getId()))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("주최자가 아닌 사람은 모임을 수정하거나 삭제할 수 없습니다.");
     }
@@ -327,7 +242,7 @@ class GroupServiceTest {
         Group savedGroup = saveGroup("모모의 스터디", Category.STUDY);
         participantService.participate(savedGroup.getId(), savedMember1.getId());
 
-        assertThatThrownBy(() -> groupService.delete(savedHost.getId(), savedGroup.getId()))
+        assertThatThrownBy(() -> groupManageService.delete(savedHost.getId(), savedGroup.getId()))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("참여자가 존재하는 모임은 수정 및 삭제할 수 없습니다.");
     }
@@ -341,7 +256,7 @@ class GroupServiceTest {
 
         long groupId = savedGroup.getId();
 
-        assertThatThrownBy(() -> groupService.delete(savedHost.getId(), groupId))
+        assertThatThrownBy(() -> groupManageService.delete(savedHost.getId(), groupId))
                 .isInstanceOf(MomoException.class)
                 .hasMessage("참여자가 존재하는 모임은 수정 및 삭제할 수 없습니다.");
     }
